Add titles to unlabelled inputs in dropdown checkbox
The dropdowns listing columns to hide or pin contain input elements
(‘Select all’ and ‘Search’) with no textual labels. The function of
these input elements may not be apparent to users of assistive
technologies. As there is no room for a visible `<label>` element, add
descriptive `title` attributes to these fields.