Omit href attributes in slider handle elements
jQuery UI 1.10 by default uses `<a href="#">` elements to represent the
handles of sliders, which may confuse assistive technologies into
describing them as hyperlinks. Providing non-default values without the
`href` attribute solves this.

Note that newer versions of jQuery UI use `<span>`s as the default value
and do not have this issue, but version 1.10 requires them to be `<a>`s